requirements.txt
astroid==2.0.4
dj-database-url==0.5.0
Django==2.1.2
django-ckeditor==5.6.1
django-heroku==0.3.1
django-js-asset==1.1.0
gunicorn==19.9.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
Pillow==5.3.0
psycopg2==2.7.5
pylint==2.1.1
pytz==2018.5
six==1.11.0
typed-ast==1.1.0
whitenoise==4.1
wrapt==1.10.11
